update positions of unit tests for resource types
