+++ koku/api/resource_types/test/Azure/__init__.py
