+++ koku/api/resource_types/test/Azure/azure_regions/__init__.py
