+++ koku/api/resource_types/test/Azure/azure_services/__init__.py
@@ -1,0 +1,1 @@
+# noqa
