+++ koku/api/resource_types/test/Azure/azure_subscription_guids/__init__.py
@@ -1,0 +1,1 @@
+# noqa
