+++ koku/api/resource_types/test/Gcp/gcp_accounts/__init__.py
@@ -1,0 +1,1 @@
+# noqa
