+++ koku/api/resource_types/test/Gcp/gcp_projects/__init__.py
@@ -1,0 +1,1 @@
+# noqa
