+++ koku/api/resource_types/test/Gcp/gcp_regions/__init__.py
@@ -1,0 +1,1 @@
+# noqa
