+++ koku/api/resource_types/test/Openshift/__init__.py
