+++ koku/api/resource_types/test/Openshift/openshift_clusters/__init__.py
@@ -1,0 +1,1 @@
+# noqa
