+++ koku/api/resource_types/test/Openshift/openshift_nodes/__init__.py
@@ -1,0 +1,1 @@
+# noqa
