+++ koku/api/resource_types/test/Openshift/openshift_projects/__init__.py
@@ -1,0 +1,1 @@
+# noqa
